[Refactor] #206 정적 팩토리 메서드명을 from으로 통일 및 패키지 구조 수정
