+++ src/main/java/org/sopt/bofit/domain/insurance/dto/response/option/RefundTypeResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.bofit.domain.insurance.dto.response;
+package org.sopt.bofit.domain.insurance.dto.response.option;
 
 import org.sopt.bofit.domain.insurance.entity.product.constant.RefundType;
 
@@ -7,7 +7,7 @@ public record RefundTypeResponse(
     String displayName
 ) {
 
-    public static RefundTypeResponse create(RefundType refundType) {
+    public static RefundTypeResponse from(RefundType refundType) {
         return new RefundTypeResponse(refundType, refundType.getDisplayName());
     }
