+++ src/main/java/org/sopt/bofit/domain/insurance/dto/response/option/RenewableTypeResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.bofit.domain.insurance.dto.response;
+package org.sopt.bofit.domain.insurance.dto.response.option;
 
 import org.sopt.bofit.domain.insurance.entity.product.constant.RenewableType;
 
@@ -7,7 +7,7 @@ public record RenewableTypeResponse(
     String displayName
 ) {
 
-    public static RenewableTypeResponse create(RenewableType renewableType){
+    public static RenewableTypeResponse from(RenewableType renewableType){
         return new RenewableTypeResponse(renewableType, renewableType.getDisplayName());
     }
